Rework plugin to make it extensible
