+++ src/controllers/OrdersController.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Commerce Insights Components Orders Controller
+ *
+ * @link      https://fostercommerce.com
+ * @copyright Copyright (c) 2021 Foster Commerce
+ */
+
+declare(strict_types = 1);
+
+namespace fostercommerce\commerceinsights\controllers;
+
+use fostercommerce\commerceinsights\CommerceInsights;
+use fostercommerce\commerceinsights\components\Orders;
+
+class OrdersController extends Orders
+{
+}
